// src/app/models/timer-row.model.ts
export class TimerRowModel{
  baseId: number;
  id: string;
  name: string;
  parentTimerLength: number;
  timerLength: number;
  timerStart: number;
  timerLengthDisplay: string;
  animationEntrance: string;
  animationExit: string;
  baseSeconds: number;
  remainingTime: string;
  timerCss: string;
  timeToFire: string;
  validForm: boolean;
  timerForm: any;
}